Sul23 611  Improve and simplify toggles (#210)
